fix(SP-7 T5): use uint64_t for R_max to prevent overflow in large scenes
N=200k gaussians with 240x135 tiles = 6.48B pairs, exceeding uint32_t
max of 4.29B. Change bin_R_max_ field and prepare_for_bin() R_max param
to uint64_t; compute R_max_estimate with uint64_t arithmetic in bin().
VkDeviceSize is already uint64_t so buffer sizing is unaffected.

+++ harmonyos_3dgs/include/vulkan/tile_binner_vulkan.h
@@ -88,7 +88,7 @@ private:
     // R_max is computed as N * num_tiles_x * num_tiles_y (safe upper bound since
     // sum(tiles_touched[i]) <= N * num_tiles by definition).
     uint32_t bin_N_     = 0u;  // Gaussian count at last alloc
-    uint32_t bin_R_max_ = 0u;  // R upper bound at last alloc
+    uint64_t bin_R_max_ = 0u;  // R upper bound at last alloc (uint64 to prevent overflow for large N * num_tiles)
     uint32_t bin_wg_    = 0u;  // workgroup count = ceil(N/256) at last alloc
 
     std::unique_ptr<VulkanBuffer> bin_tt_buf_;    // tiles_touched [N] i32
@@ -101,5 +101,5 @@ private:
     std::unique_ptr<VulkanBuffer> bin_keys_buf_;  // keys_unsorted [R_max] u64
     std::unique_ptr<VulkanBuffer> bin_vals_buf_;  // values_unsorted [R_max] u32
 
-    void prepare_for_bin(uint32_t N, uint32_t R_max, uint32_t num_wgs);
+    void prepare_for_bin(uint32_t N, uint64_t R_max, uint32_t num_wgs);
 };
